home category listenin itemi ayrıldı.
